+++ extensions/amazon-aws/src/components/common/action.tsx
@@ -1,14 +1,16 @@
-import { Action, ActionPanel } from "@raycast/api";
+import { Action, ActionPanel, Icon, Keyboard } from "@raycast/api";
+import { getEnumKeysExcludingCurrent } from "../../util";
 
 export class AwsAction {
-  public static Console = ({ url }: { url: string }) => (
+  public static Console = ({ url, onAction }: { url: string; onAction?: () => void }) => (
     <ActionPanel.Section title="Console">
-      <Action.OpenInBrowser key="openConsoleLink" title="Open in Browser" url={url} />
+      <Action.OpenInBrowser key="openConsoleLink" title="Open in Browser" url={url} onOpen={() => onAction?.()} />
       <Action.CopyToClipboard
         key="copyConsoleLink"
         title="Copy Link"
         content={url}
-        shortcut={{ modifiers: ["cmd", "shift"], key: "c" }}
+        shortcut={Keyboard.Shortcut.Common.Copy}
+        onCopy={() => onAction?.()}
       />
     </ActionPanel.Section>
   );
@@ -20,4 +22,35 @@ export class AwsAction {
       shortcut={{ modifiers: ["opt"], key: "e" }}
     />
   );
+
+  public static SwitchResourceType<T extends object>({
+    setResourceType,
+    current,
+    enumType,
+  }: {
+    setResourceType: (resourceType: T[keyof T]) => void;
+    current: T[keyof T];
+    enumType: T;
+  }) {
+    const allKeysExceptCurrent = getEnumKeysExcludingCurrent(enumType, current);
+    return (
+      <ActionPanel.Section>
+        <ActionPanel.Submenu
+          title="Switch Resource Type"
+          shortcut={{ modifiers: ["cmd", "shift"], key: "r" }}
+          filtering
+          icon={Icon.Box}
+        >
+          {allKeysExceptCurrent.map((k) => (
+            <Action
+              title={`Show ${k.toString()}`}
+              onAction={() => setResourceType(enumType[k])}
+              key={k.toString()}
+              icon={Icon.Box}
+            />
+          ))}
+        </ActionPanel.Submenu>
+      </ActionPanel.Section>
+    );
+  }
 }