+++ extensions/cal-com-share-meeting-links/CHANGELOG.md
@@ -1,10 +1,19 @@
 # Cal.com Share Meeting Links Changelog
 
-## 2023-12-04
+## [Updated] - 2024-06-12
+
+- Adds submenu to View Bookings command to enable you to update a booking's status
+
+## [Update] - 2024-06-04
+
+- Adds View Bookings command, which displays the user's bookings
+- View Bookings includes an action to enable you to cancel bookings
+
+## [Update] - 2023-12-04
 
 - Sets the event type's length as a keyword so that you can still search for "60" to find an event type named "1 hour call"
 
-## 2023-09-10
+## [Update] - 2023-09-10
 
 - Sorts event types to match their order in the dashboard
 - Adds actions to open your dashboard and copy the link to your public page